add make_dictionary
This is a separate file for making the dictionary.json file from the organized.txt file. This is occasionally used in the case that some word is missing from a few spots and can be manually added to organized.txt. This prevents spending a hour waiting for organize2.py to redo the entire dictionary needlessly.