added to support for dynamically adding keys
I still need to test this, but this update is meant to finish implementing addKey and changes to main so that if the user input is not initially found, it will search the whole
english dictionary using that key to double check if there are no words that can be made. My initial structure was flawed and means queries like 'berryz' would not return any words that could be made. If any words can be made from the key it will be added to the dictionary for future reference.